PHASE 2
Environmental monitoring refers to the tools and techniques designed to observe an environment, characterize its quality, and establish environmental parameters, for the purpose of accurately quantifying the impact an activity has on an environment. Results are gathered, analyzed statistically, and then published in a risk assessment and environmental monitoring and impact assessment report.